requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
fsspec==2025.3.2
googlesearch-python==1.3.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==80.0.0
shapely==2.1.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tifffile==2025.3.30
tldextract==5.3.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.119
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
